requirements.txt
fastapi
uvicorn
pdfplumber
sentence-transformers
faiss-cpu
python-multipart
aiofiles
openai
dotenv